requirement.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
python-dotenv==1.0.1
slowapi==0.1.9
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2